Fix #5009 - Refactor AbstractMenu/Splitbutton common behaviors
