+++ src/main/java/org/primefaces/component/splitbutton/SplitButton.java
@@ -23,27 +23,16 @@
  */
 package org.primefaces.component.splitbutton;
 
-import java.util.List;
-import javax.el.ELContext;
-import javax.el.ExpressionFactory;
-import javax.el.MethodExpression;
-import javax.el.MethodNotFoundException;
-import javax.faces.application.ResourceDependencies;
-import javax.faces.application.ResourceDependency;
-import javax.faces.context.FacesContext;
-import javax.faces.event.AbortProcessingException;
-import javax.faces.event.ActionEvent;
-import javax.faces.event.FacesEvent;
-import org.primefaces.event.MenuActionEvent;
-import org.primefaces.model.menu.MenuElement;
-import org.primefaces.model.menu.MenuGroup;
-import org.primefaces.model.menu.MenuItem;
 import org.primefaces.model.menu.MenuModel;
-
 import org.primefaces.util.ComponentUtils;
 import org.primefaces.util.HTML;
 import org.primefaces.util.LangUtils;
-import org.primefaces.util.SerializableFunction;
+
+import javax.faces.application.ResourceDependencies;
+import javax.faces.application.ResourceDependency;
+import javax.faces.event.AbortProcessingException;
+import javax.faces.event.FacesEvent;
+import java.util.List;
 
 @ResourceDependencies({
         @ResourceDependency(library = "primefaces", name = "components.css"),
@@ -132,80 +121,8 @@ public class SplitButton extends SplitButtonBase {
         return (elements == null) ? 0 : elements.size();
     }
 
-    protected MenuItem findMenuitem(List<MenuElement> elements, String id) {
-        if (elements == null || elements.isEmpty()) {
-            return null;
-        }
-        else {
-            String[] paths = id.split("_");
-
-            if (paths.length == 0) {
-                return null;
-            }
-
-            int childIndex = Integer.parseInt(paths[0]);
-            if (childIndex >= elements.size()) {
-                return null;
-            }
-
-            MenuElement childElement = elements.get(childIndex);
-
-            if (paths.length == 1) {
-                return (MenuItem) childElement;
-            }
-            else {
-                String relativeIndex = id.substring(id.indexOf('_') + 1);
-
-                return findMenuitem(((MenuGroup) childElement).getElements(), relativeIndex);
-            }
-        }
-    }
-
     @Override
     public void broadcast(FacesEvent event) throws AbortProcessingException {
-        if (event instanceof MenuActionEvent) {
-            FacesContext facesContext = getFacesContext();
-
-            MenuActionEvent menuActionEvent = (MenuActionEvent) event;
-            MenuItem menuItem = menuActionEvent.getMenuItem();
-
-            SerializableFunction<MenuItem, String> function = menuItem.getFunction();
-            String command = menuItem.getCommand();
-            if (function != null) {
-                String outcome = function.apply(menuItem);
-                facesContext.getApplication().getNavigationHandler().handleNavigation(facesContext, null, outcome);
-            }
-            else if (command != null) {
-                ELContext elContext = facesContext.getELContext();
-                ExpressionFactory expressionFactory = facesContext.getApplication().getExpressionFactory();
-
-                Object invokeResult = null;
-                try {
-                    MethodExpression me = expressionFactory.createMethodExpression(elContext, command,
-                                String.class, new Class[0]);
-                    invokeResult = me.invoke(elContext, null);
-                }
-                catch (MethodNotFoundException mnfe1) {
-                    try {
-                        MethodExpression me = expressionFactory.createMethodExpression(elContext, command,
-                                        String.class, new Class[]{ActionEvent.class});
-                        invokeResult = me.invoke(elContext, new Object[]{event});
-                    }
-                    catch (MethodNotFoundException mnfe2) {
-                        MethodExpression me = expressionFactory.createMethodExpression(elContext, command,
-                                        String.class, new Class[]{MenuActionEvent.class});
-                        invokeResult = me.invoke(elContext, new Object[]{event});
-                    }
-                }
-                finally {
-                    String outcome = (invokeResult != null) ? invokeResult.toString() : null;
-
-                    facesContext.getApplication().getNavigationHandler().handleNavigation(facesContext, command, outcome);
-                }
-            }
-        }
-        else {
-            super.broadcast(event);
-        }
+        handleBroadcast(event, getFacesContext(), super::broadcast);
     }
 }
