+++ src/main/java/org/primefaces/renderkit/OutcomeTargetRenderer.java
@@ -23,27 +23,17 @@
  */
 package org.primefaces.renderkit;
 
-import org.primefaces.behavior.confirm.ConfirmBehavior;
-import org.primefaces.component.api.AjaxSource;
 import org.primefaces.component.api.UIOutcomeTarget;
 import org.primefaces.context.PrimeApplicationContext;
-import org.primefaces.model.menu.MenuItem;
-import org.primefaces.util.ComponentTraversalUtils;
 
 import javax.faces.FacesException;
 import javax.faces.application.ConfigurableNavigationHandler;
 import javax.faces.application.NavigationCase;
 import javax.faces.component.UIComponent;
-import javax.faces.component.UIForm;
-import javax.faces.component.behavior.ClientBehavior;
-import javax.faces.component.behavior.ClientBehaviorContext;
-import javax.faces.component.behavior.ClientBehaviorHolder;
 import javax.faces.context.FacesContext;
-import javax.faces.context.ResponseWriter;
 import javax.faces.event.ActionListener;
 import javax.faces.flow.FlowHandler;
 import javax.faces.lifecycle.ClientWindow;
-import java.io.IOException;
 import java.util.*;
 import org.primefaces.component.api.ClientBehaviorRenderingMode;
 import org.primefaces.component.api.DialogReturnAware;
@@ -213,92 +203,5 @@ public class OutcomeTargetRenderer extends CoreRenderer {
         }
 
         return url;
-    }
-
-    protected void encodeOnClick(FacesContext context, UIComponent source, MenuItem menuitem) throws IOException {
-        ResponseWriter writer = context.getResponseWriter();
-        setConfirmationScript(context, menuitem);
-
-        String onclick = menuitem.getOnclick();
-
-        //GET
-        if (menuitem.getUrl() != null || menuitem.getOutcome() != null) {
-            String targetURL = getTargetURL(context, (UIOutcomeTarget) menuitem);
-            writer.writeAttribute("href", targetURL, null);
-
-            if (menuitem.getTarget() != null) {
-                writer.writeAttribute("target", menuitem.getTarget(), null);
-            }
-        }
-        //POST
-        else {
-            writer.writeAttribute("href", "#", null);
-
-            UIForm form = ComponentTraversalUtils.closestForm(context, source);
-            if (form == null) {
-                throw new FacesException("MenuItem must be inside a form element");
-            }
-
-            String command;
-            if (menuitem.isDynamic()) {
-                String menuClientId = source.getClientId(context);
-                Map<String, List<String>> params = menuitem.getParams();
-                if (params == null) {
-                    params = new LinkedHashMap<>();
-                }
-                List<String> idParams = Collections.singletonList(menuitem.getId());
-                params.put(menuClientId + "_menuid", idParams);
-
-                command = menuitem.isAjax()
-                        ? buildAjaxRequest(context, source, (AjaxSource) menuitem, form, params)
-                        : buildNonAjaxRequest(context, source, form, menuClientId, params, true);
-            }
-            else {
-                command = menuitem.isAjax()
-                        ? buildAjaxRequest(context, (UIComponent & AjaxSource) menuitem, form)
-                        : buildNonAjaxRequest(context, ((UIComponent) menuitem), form, ((UIComponent) menuitem).getClientId(context), true);
-            }
-
-            onclick = (onclick == null) ? command : onclick + ";" + command;
-        }
-
-        if (onclick != null) {
-            if (menuitem.requiresConfirmation()) {
-                writer.writeAttribute("data-pfconfirmcommand", onclick, null);
-                writer.writeAttribute("onclick", menuitem.getConfirmationScript(), "onclick");
-            }
-            else {
-                writer.writeAttribute("onclick", onclick, null);
-            }
-        }
-
-        if (menuitem instanceof DialogReturnAware) {
-            List<ClientBehaviorContext.Parameter> behaviorParams = new ArrayList<>();
-            behaviorParams.add(new ClientBehaviorContext.Parameter(Constants.CLIENT_BEHAVIOR_RENDERING_MODE, ClientBehaviorRenderingMode.UNOBSTRUSIVE));
-            String dialogReturnBehavior = getEventBehaviors(context, (ClientBehaviorHolder) menuitem, DialogReturnAware.EVENT_DIALOG_RETURN,
-                    behaviorParams);
-            if (dialogReturnBehavior != null) {
-                writer.writeAttribute(DialogReturnAware.ATTRIBUTE_DIALOG_RETURN_SCRIPT, dialogReturnBehavior, null);
-            }
-        }
-    }
-
-    protected void setConfirmationScript(FacesContext context, MenuItem item) {
-        if (item instanceof ClientBehaviorHolder) {
-            Map<String, List<ClientBehavior>> behaviors = ((ClientBehaviorHolder) item).getClientBehaviors();
-            List<ClientBehavior> clickBehaviors = (behaviors == null) ? null : behaviors.get("click");
-
-            if (clickBehaviors != null && !clickBehaviors.isEmpty()) {
-                for (int i = 0; i < clickBehaviors.size(); i++) {
-                    ClientBehavior clientBehavior = clickBehaviors.get(i);
-                    if (clientBehavior instanceof ConfirmBehavior) {
-                        ClientBehaviorContext cbc = ClientBehaviorContext.createClientBehaviorContext(
-                                context, (UIComponent) item, "click", item.getClientId(), Collections.EMPTY_LIST);
-                        clientBehavior.getScript(cbc);
-                        break;
-                    }
-                }
-            }
-        }
     }
 }
